pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-document-system"
version = "0.1.0"
description = "Multi-AI-Agent Document System for Job–Resume Matching"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "scipy>=1.11.0",
    "rank-bm25>=0.2.2",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
doc-system = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
